+++ src/models/quizAttempt.model.js
@@ -3,7 +3,7 @@ const sequelize = require('../config/database');
 
 const QuizAttempt = sequelize.define('QuizAttempt', {
   quiz_attempt_id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  child_profile_id: { type: DataTypes.INTEGER, allowNull: false },
+  child_profile_id: { type: DataTypes.BIGINT, allowNull: false },
   quiz_set_id: { type: DataTypes.INTEGER, allowNull: false },
   total_questions: { type: DataTypes.INTEGER, allowNull: false },
   correct_count: { type: DataTypes.INTEGER, allowNull: false },
