+++ src/models/vocabularyEntry.model.js
@@ -5,8 +5,8 @@ const VocabularyEntry = sequelize.define(
   'VocabularyEntry',
   {
     vocabulary_entry_id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-    child_profile_id: { type: DataTypes.INTEGER, allowNull: false },
-    story_id: { type: DataTypes.INTEGER, allowNull: true },
+    child_profile_id: { type: DataTypes.BIGINT, allowNull: false },
+    story_id: { type: DataTypes.BIGINT, allowNull: true },
     english_word: { type: DataTypes.STRING(100), allowNull: false },
     korean_meaning: { type: DataTypes.STRING(255), allowNull: false },
     example_sentence: { type: DataTypes.TEXT, allowNull: true },
